+++ Samples/C#/SimpleConsoleApplication/.gitignore
