+++ Samples/C#/SimpleConsoleApplication/Program.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace SimpleConsoleApplication
+{
+	class Program
+	{
+		static void Main(string[] args)
+		{
+			Console.WriteLine("Hello World, Soup Style!");
+		}
+	}
+}
