+++ Samples/C#/SimpleConsoleApplication/Recipe.toml
@@ -1,0 +1,7 @@
+Name = "SimpleConsoleApplication"
+Language = "C#"
+Type = "Executable"
+Version = "1.1.3"
+Source = [
+	"Program.cs"
+]
