+++ Samples/C++/ModuleConsoleApplication/.gitignore
