+++ Samples/C++/ModuleConsoleApplication/Main.cpp
