+++ Samples/C++/ModuleConsoleApplication/Module.cpp
