+++ Samples/C++/ParseJsonFile/Message.json
@@ -1,0 +1,3 @@
+{
+  "message": "Hello!"
+}
