+++ Samples/C++/ParseJsonFile/Recipe.toml
@@ -1,0 +1,10 @@
+Name = "ParseJsonFile"
+Language = "C++"
+Version = "1.0.0"
+Type = "Executable"
+Dependencies = [
+	"json11@1.0.1",
+]
+Source = [
+	"Main.cpp",
+]
