+++ Samples/C++/SimpleBuildExtension/.gitignore
