+++ Samples/C++/SimpleBuildExtension/Extension/Register.cpp
