+++ Samples/C++/SimpleConsoleApplication/.gitignore
