+++ Samples/C++/SimpleConsoleApplication/Main.cpp
