+++ Samples/C++/StaticLibraryReference/MyApplication/Main.cpp
