+++ Samples/C++/StaticLibraryReference/MyApplication/Recipe.toml
