+++ Samples/C++/StaticLibraryReference/MyLibrary/Module.cpp
