+++ Samples/C++/StaticLibraryReference/MyLibrary/Recipe.toml
