--- Samples/ParseJsonFile/Message.json
@@ -1,3 +1,0 @@
-{
-  "message": "Hello!"
-}
