--- Samples/ParseJsonFile/Recipe.toml
@@ -1,10 +1,0 @@
-Name = "ParseJsonFile"
-Language = "C++"
-Version = "1.0.0"
-Type = "Executable"
-Dependencies = [
-	"json11@1.0.1",
-]
-Source = [
-	"Main.cpp",
-]
